weather-service/requirements.txt
grpcio==1.75.0
grpcio-tools==1.75.0
requests==2.32.5
